+++ aim/sdk/artifacts/serializable.py
@@ -12,9 +12,14 @@ class Serializable(metaclass=ABCMeta):
     """
     # List of types in which passed serialized content can be saved
     JSON_FILE = 'json_file'
+    LOG_FILE = 'log_file'
     DIR = 'dir'
     IMAGE = 'image'
     MODEL = 'model'
+
+    # List of available file extensions
+    JSON_EXT = 'json'
+    LOG_EXT = 'log'
 
     # List of different modes of storing serialized content to a file
     CONTENT_MODE_WRITE = 'w'
@@ -33,11 +38,14 @@ class Serializable(metaclass=ABCMeta):
         ...
 
     @staticmethod
-    def store_json(repo, content, dir_path=None) -> dict:
-        file_name = '{}.json'.format(content['name'])
+    def store_file(repo, ext, content, dir_path=None) -> dict:
         data = content['data'] if 'data' in content else {}
+        file_name = '{n}.{e}'.format(n=content['name'],
+                                     e=ext)
 
         res = repo.store_file(file_name,
+                              ext,
+                              content['name'],
                               content['cat'],
                               content['content'],
                               content['mode'],
@@ -55,7 +63,10 @@ class Serializable(metaclass=ABCMeta):
         for stored_obj, content in list(serialized_inst.items()):
             if stored_obj == self.JSON_FILE:
                 # Store artifact inside json file
-                res = self.store_json(repo, content)
+                res = self.store_file(repo, self.JSON_EXT, content)
+            elif stored_obj == self.LOG_FILE:
+                # Store artifact as log file
+                res = self.store_file(repo, self.LOG_EXT, content)
             elif stored_obj == self.DIR:
                 dir_path, dir_rel_path = repo.store_dir(content['name'],
                                                         content['data'])
@@ -63,7 +74,9 @@ class Serializable(metaclass=ABCMeta):
                 for f in content['files']:
                     # Store dir files
                     for _, f_content in f.items():
-                        res.append(self.store_json(repo,
+                        res.append(self.store_file(repo,
+                                                   # TODO: get ext from content
+                                                   'log',
                                                    f_content,
                                                    dir_rel_path))
             elif stored_obj == self.IMAGE: