+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Update tracking artifacts to be saved in log format (gorarakelyan)
 - Add pytorch cuda support to existing sdk artefacts (gorarakelyan)
 - Add cli reset command (gorarakelyan)
 - Add nested module tracking support to aim sdk (gorarakelyan)